Applied patch #6677
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@27080 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSProxy.m
@@ -275,7 +275,7 @@ extern BOOL __objc_responds_to(id, SEL);
 - (NSString*) description
 {
   return [NSString stringWithFormat: @"<%s %lx>",
-	GSClassNameFromObject(self), (unsigned long)self];
+	GSClassNameFromObject(self), (size_t)self];
 }
 
 /**
